core/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Settings:
    TITLE: str = "Gen API"
    DATABASE_URL: str = (
        f"postgresql://{os.getenv('DB_USER')}:{os.getenv('DB_PASSWORD')}@{os.getenv('DB_HOST')}:{os.getenv('DB_PORT')}/{os.getenv('DB_NAME')}"
    )


settings: Settings = Settings()
